backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1

# Data
yfinance==0.2.36
pandas==2.2.0
numpy==1.26.4
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0

# ML / Deep Learning
torch==2.2.0
transformers==4.38.1
scikit-learn==1.4.0

# RL
gymnasium==0.29.1

# Portfolio Optimization
scipy==1.12.0
cvxpy==1.4.2

# Explainability
shap==0.44.1
captum==0.7.0

# Visualization / API support
plotly==5.18.0
matplotlib==3.8.3

# Utilities
joblib==1.3.2
tqdm==4.66.2
loguru==0.7.2
